fix format strings

+++ fpga/lib/ips/rtds2gpu/gpu2rtds.cpp
@@ -48,9 +48,9 @@ void Gpu2Rtds::dump(spdlog::level::level_enum logLevel)
 	logger->log(logLevel, "    Frame too short:    {}", (status.frame_too_short ? "yes" : "no"));
 	logger->log(logLevel, "    Frame too long:     {}", (status.frame_too_long ? "yes" : "no"));
 	logger->log(logLevel, "    Frame size invalid: {}", (status.invalid_frame_size ? "yes" : "no"));
-	logger->log(logLevel, "    Last count:         {}", status.last_count);
-	logger->log(logLevel, "    Last seq. number:   {}", status.last_seq_nr);
-	logger->log(logLevel, "    Max. frame size:    {}", status.max_frame_size);
+	logger->log(logLevel, "    Last count:         {}", (int) status.last_count);
+	logger->log(logLevel, "    Last seq. number:   {}", (int) status.last_seq_nr);
+	logger->log(logLevel, "    Max. frame size:    {}", (int) status.max_frame_size);
 }
 
 //bool Gpu2Rtds::startOnce(const MemoryBlock &mem, size_t frameSize, size_t dataOffset, size_t doorbellOffset)
@@ -126,6 +126,6 @@ Gpu2Rtds::getMaxFrameSize()
 //	logger->info("Doorbell register: {:#08x}", doorbell.value);
 //	logger->info("  Valid:       {}", (doorbell.is_valid ? "yes" : "no"));
 //	logger->info("  Count:       {}", doorbell.count);
-//	logger->info("  Seq. number: {}", doorbell.seq_nr);
+//	logger->info("  Seq. number: {}", (int) doorbell.seq_nr);
 //}
